refactor: update input validation logic
InputValidator now deals with the validation of user input
